Run task on user-supplied cron schedule
Keeping it inside the code instead of setting a cron schedule in the OS
allows one built Docker image to be configured according to users schedules.

If it was in the OS, the Dockerfile would likely have to setup that cron
schedule as part of the build process, meaning each config change needs
a new build.